refactor: DRY - abstracted copy to clipboard
